+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Support for more API methods in Voucher API
 
 ### Changed
 
-*
+* Bumped dependencies
 
 ### Fixed